Resolving libdl symbols
It turned out that we will actually need to use proper libdl and
libpthread libraries for injection. That's why we are going to look for
the real libraries.

Also, add support for ELFOSABINONE, libdl on my machine seems to be
using this ABI in its ELF header. It still parses okay.