Remote clone() syscall
Add support for injecting clone() system calls into remote processes.

clone() actually takes a bunch of optional pointer arguments, but we
will not need to use them all.

In order to be able to trace the newly created process we also set the
PTRACE_O_TRACECLONE option. The thread will be created pre-stopped and
we will need to restart it later. The parent tracer will also receive a
PTRACE_EVENT_CLONE which we need to skip over.  Without this option the
new thread will be untraceable and unwaitable.