Clear ptrace options in the helper thread
clone() duplicates ptrace options as well. We need to clear them so that
our wait() does not stop on the syscalls performed by the new thread.
In particular, we are not interested in the PTRACE_EVENT_CLONE produced
by pthread_create(). We need only the exit code of the thread.