Ensure 16-byte alignment of stack
This is because %rsp has to be aligned to 16 bytes [1]. Otherwise
dlopen() or pthread_create() may SIGSEGV when spilling XMM registers
onto the stack (XMM requires alignment and GCC expects that).

Note that we write the return address for clone at 8-byte alignment
which will realign us back to "base + 1 MB" when "retq" is executed.
However, *we* need to ensure that %rsp stays aligned when calling.

[1]: https://sourceforge.net/p/fbc/bugs/659/